+++ Contents/mods/pz-zombie-factions/42/media/lua/client/ZombieFactions/ImpactProbe_Client.lua
@@ -33,7 +33,7 @@ local processedOwnerHits = {}
 local processedOwnerHitOrder = {}
 local activeHitReactions = {}
 
-print("[ZombieFactions] Client impact probe loaded v0.0.33")
+print("[ZombieFactions] Client impact probe loaded v0.0.34")
 
 local function print(message)
     CombatController.detail(message)